[DWRF] bug casts all integers to short from the stride dictionaries
Summary: a left over cast, casts all numeric values from stride dictionaries to short causing wrong results

Test Plan: adjusted TestOrcFile so values in int and long columns can only be int or long, this fails before the fix

+++ VENDOR.hive/trunk/ql/src/java/com/facebook/hive/orc/lazy/LazyNumericDictionaryTreeReader.java
@@ -104,7 +104,7 @@ abstract class LazyNumericDictionaryTreeReader extends LazyTreeReader {
     if (isInDictionary) {
       return dictionaryValues[(int) reader.next()];
     } else {
-      return (short) reader.next();
+      return reader.next();
     }
   }
 }
